fix(probes): isolate global-subscriber tests to kill #203 pollution flake
`install_probe_tracing` calls `tracing_subscriber::registry().try_init()`,
which permanently attaches the substrate's UriCaptureLayer +
ProbeRouterLayer + fmt layer to the GLOBAL tracing default for the
rest of the process. That's correct for production (one process, one
boot, one install) but in cargo's lib-test binary it leaked across
tests:

  routing::tracing_init::tests::install_is_idempotent_with_no_disk_capture
    → installs the global subscriber

  routing::uri_layer::tests::no_subscriber_returns_empty_chain
    → asserts an empty chain "no installed Layer means no captured frames"
    → PANICS when the install ran first in the same binary

Task #203 captured the symptom. Today's PR #1547 hit it deterministically
when running `cargo test -p continuum-core --lib routing::`.

Fix: move both `install_*` tests out of the lib-test binary and into
their own integration-test file (`tests/tracing_install_global.rs`).
Each `tests/*.rs` file becomes a separate binary in cargo's runner —
process isolation contains the global install per binary.

`from_env_reads_documented_env_vars` stays in the lib-test module
because it only touches env vars (already serial within itself) and
never calls `install_probe_tracing`.

Verification: `cargo test -p continuum-core --lib routing::` now
passes 252/252 (was 252/253 with the flake), and the new integration
binary passes 2/2.
